исправлен скрипты создания таблиц, реализована логика получения достижений

+++ JDBC/telegram_pomodoro-main/src/main/resources/migrations/V1__create_table_user_sessions.sql
@@ -1,7 +1,7 @@
 CREATE TABLE user_sessions
 (
     id        SERIAL primary key,
-    chat_id   int         not null,
+    chat_id   bigint        not null,
     type      varchar(10) not null,
     duration  int,
     start_at  timestamp   not null,
